TLS - Prevent Duplicate Entries in .env File
Previously, when generating local certificates, running the command multiple times could result in duplicate lines being written to the `.env` file. This commit resolves the issue by implementing the same mechanism used in the Let's Encrypt commands, ensuring that duplicate entries are avoided.

(cherry picked from commit 1aa95087765aad0c0e1c91bfe2158afa2b12cc9b)
